mysql-apt

+++ plugins/mariadb/index.py
@@ -335,6 +335,10 @@ def initMysqlData():
         data = mw.execShell(cmd)
         # print(data[0])
         # print(data[1])
+
+        if not mw.isAppleSystem():
+            mw.execShell('chown -R mysql:mysql ' + serverdir + '/data')
+            mw.execShell('chmod -R 755 ' + serverdir + '/data')
         return False
     return True
 
@@ -359,6 +363,11 @@ def initMariaDbPwd():
         print("init mariadb password fail:" + data[1])
         exit(1)
 
+    # 删除空账户
+    drop_empty_user = serverdir + '/bin/mysql ' + db_option + '-uroot -p' + \
+        pwd + ' -e "use mysql;delete from user where USER=\'\'"'
+    mw.execShell(drop_empty_user)
+
     # 删除测试数据库
     drop_test_db = serverdir + '/bin/mysql ' + db_option + ' -uroot -p' + \
         pwd + ' -e "drop database test";'
@@ -366,13 +375,14 @@ def initMariaDbPwd():
 
     # 删除冗余账户
     hostname = mw.execShell('hostname')[0].strip()
-    drop_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
-        myconf + ' -uroot -p' + pwd + ' -e "drop user \'\'@\'' + hostname + '\'";'
-    mw.execShell(drop_hostname)
+    if hostname != 'localhost':
+        drop_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
+            myconf + ' -uroot -p' + pwd + ' -e "drop user \'\'@\'' + hostname + '\'";'
+        mw.execShell(drop_hostname)
 
-    drop_root_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
-        myconf + ' -uroot -p' + pwd + ' -e "drop user \'root\'@\'' + hostname + '\'";'
-    mw.execShell(drop_root_hostname)
+        drop_root_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
+            myconf + ' -uroot -p' + pwd + ' -e "drop user \'root\'@\'' + hostname + '\'";'
+        mw.execShell(drop_root_hostname)
 
     pSqliteDb('config').where('id=?', (1,)).save('mysql_root', (pwd,))
     return True
