+++ plugins/mysql/index.py
@@ -407,7 +407,7 @@ def initMysqlPwd():
     myconf = serverdir + "/etc/my.cnf"
     pwd = mw.getRandomString(16)
 
-    cmd_pass = serverdir + '/bin/mysql -uroot -e'
+    cmd_pass = serverdir + '/bin/mysql --defaults-file=' + myconf + ' -uroot -e'
     cmd_pass = cmd_pass + \
         '"UPDATE mysql.user SET password=PASSWORD(\'' + \
         pwd + "') WHERE user='root';"
@@ -428,14 +428,14 @@ def initMysqlPwd():
 
     # 删除冗余账户
     hostname = mw.execShell('hostname')[0].strip()
+    if hostname != 'localhost':
+        drop_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
+            myconf + ' -uroot -p' + pwd + ' -e "drop user \'\'@\'' + hostname + '\'";'
+        mw.execShell(drop_hostname)
 
-    drop_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
-        myconf + ' -uroot -p' + pwd + ' -e "drop user \'\'@\'' + hostname + '\'";'
-    mw.execShell(drop_hostname)
-
-    drop_root_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
-        myconf + ' -uroot -p' + pwd + ' -e "drop user \'root\'@\'' + hostname + '\'";'
-    mw.execShell(drop_root_hostname)
+        drop_root_hostname =  serverdir + '/bin/mysql  --defaults-file=' + \
+            myconf + ' -uroot -p' + pwd + ' -e "drop user \'root\'@\'' + hostname + '\'";'
+        mw.execShell(drop_root_hostname)
 
     pSqliteDb('config').where('id=?', (1,)).save('mysql_root', (pwd,))
     return True
@@ -836,7 +836,7 @@ def setDbBackup():
 
     scDir = mw.getRunDir() + '/scripts/backup.py'
 
-    cmd = 'python ' + scDir + ' database ' + args['name'] + ' 3'
+    cmd = 'python3 ' + scDir + ' database ' + args['name'] + ' 3'
     os.system(cmd)
     return mw.returnJson(True, 'ok')
